shift Verifier trait methods into primitives
